unit test done & readme

+++ Anchor-Editor-Backend/Anchor-Editor-Backend/Controllers/AnchorsController.cs
@@ -31,7 +31,6 @@ namespace Anchor_Editor_Backend.Controllers
         [HttpGet]
         public IActionResult GetAllAnchors()
         {
-            
             return Ok(_anchorRepository.AnchorList);
         }
 
@@ -39,40 +38,95 @@ namespace Anchor_Editor_Backend.Controllers
         [HttpGet("{timestamp}")]
         public IActionResult GetAnchorsByTimestamp(string timestamp)
         {
-            Anchor anchor = _anchorRepository.GetAnchorByTimestamp(timestamp);
-            return Ok(anchor);
+            if(string.IsNullOrEmpty(timestamp))
+            {
+                return BadRequest("timestamp not valid");
+            }
+            try
+            {
+                Anchor anchor = _anchorRepository.GetAnchorByTimestamp(timestamp);
+                return Ok(anchor);
+            }
+            catch(Exception ex)
+            {
+                return BadRequest(ex.Message);
+            }
         }
 
         [EnableCors("AllowOrigin")]
         [HttpGet("{location:int}")]
         public IActionResult GetAnchorsByLocation(int location)
         {
-            IList <Anchor> anchors = _anchorRepository.GetAnchorByLocation(location);
-            return Ok(anchors);
+            if (location < 0)
+            {
+                return BadRequest("location not valid");
+            }
+            try
+            {
+                IList<Anchor> anchors = _anchorRepository.GetAnchorsByLocation(location);
+                return Ok(anchors);
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(ex.Message);
+            }
         }
 
         [EnableCors("AllowOrigin")]
         [HttpDelete("{timestamp}")]
         public IActionResult DeleteAnchorByTimestamp(string timestamp)
         {
-            _anchorRepository.DeleteAnchorByTimestamp(timestamp);
-            return Ok(_anchorRepository.AnchorList);
+            if (string.IsNullOrEmpty(timestamp))
+            {
+                return BadRequest("timestamp not valid");
+            }
+            try
+            {
+                _anchorRepository.DeleteAnchorByTimestamp(timestamp);
+                return Ok(_anchorRepository.AnchorList);
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(ex.Message);
+            }
         }
 
         [EnableCors("AllowOrigin")]
         [HttpPost]
         public IActionResult AddAnchorByTimestamp([FromQuery] string destinationTimestamp, [FromQuery] int destinationLocation)
         {
-            _anchorRepository.AddAnchorByTimestamp(destinationTimestamp, destinationLocation);
-            return Ok(_anchorRepository.AnchorList);
+            if (string.IsNullOrEmpty(destinationTimestamp))
+            {
+                return BadRequest("destinationTimestamp not valid");
+            }
+            try
+            {
+                _anchorRepository.AddAnchorByTimestamp(destinationTimestamp, destinationLocation);
+                return Ok(_anchorRepository.AnchorList);
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(ex.Message);
+            }
         }
 
         [EnableCors("AllowOrigin")]
         [HttpPut]
         public IActionResult EditAnchor([FromQuery] string originalTimestamp, [FromQuery] int originalLocation, [FromQuery] string destinationTimestamp, [FromQuery] int destinationLocation)
         {
-            _anchorRepository.EditAnchor(originalTimestamp, originalLocation, destinationTimestamp, destinationLocation);
-            return Ok(_anchorRepository.AnchorList);
+            if (string.IsNullOrEmpty(originalTimestamp))
+            {
+                return BadRequest("originalTimestamp not valid");
+            }
+            try
+            {
+                _anchorRepository.EditAnchor(originalTimestamp, originalLocation, destinationTimestamp, destinationLocation);
+                return Ok(_anchorRepository.AnchorList);
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(ex.Message);
+            }
         }
     }
 }
